修复bug（诸如1、缓存的数据没有在save后刷新，就算刷新了也无法保证直接修改存档后能改动缓存，遂直接删除存档缓存。2、村民建筑的交互逻辑，修改为绑定关系和是否在其中……还有很多）。 添加主世界村民面板，目前针对住房村民通过简单测试，实现滚动列表，下一步应实现SmoothMove以及相关的调遣功能。

+++ Assets/ExtendFrame/NSNodeGraph/Editor/ProjectNodeView.cs
@@ -13,7 +13,7 @@ namespace ExtendFrame {
 		public string Tag, Text;
 		private readonly List<NSPairRef<RepoType, ulong>> _demands = new();
 		private readonly List<NSPairRef<ArchType, int>> _unlocks = new();
-		private readonly List<NSPairRef<ArchType, int>> _buffs = new();
+		private readonly List<NSPairRef<ArchType, float>> _buffs = new();
 		public TextField TextTF, TagTF;
 
 		private VisualElement _demandFoldout, _unlockFoldout, _buffFoldout;
@@ -96,7 +96,7 @@ namespace ExtendFrame {
 		private void AddNewBuff() {
 			AddBuff(new(ArchType.Cottage, 0));
 		}
-		public void AddBuff(NSPairRef<ArchType, int> buff) {
+		public void AddBuff(NSPairRef<ArchType, float> buff) {
 			_buffs.Add(buff);
 			var buffContainer = new VisualElement();
 			buffContainer.style.flexDirection = FlexDirection.Row;
